Refactoring package names
