+++ pubsub/solace/amqp/amqp.go
