+++ pubsub/solace/amqp/metadata.go
