+++ tests/config/pubsub/solace/amqp/pubsub.yml
@@ -1,9 +1,7 @@
 apiVersion: dapr.io/v1alpha1
 kind: Component
-metadata:
-  name: solace
 spec:
-  type: pubsub.amqp
+  type: pubsub.solace.amqp
   version: v1
   metadata:
     - name: url
@@ -14,4 +12,3 @@ spec:
       value: 'default'
     - name: anonymous
       value: true
-
